+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 
 * Refactoring CLI modules into a single package, rather than by CLI group
   [#81](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/81)
+* Refactoring database functions from CLI module to database client class
+  [#71](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/71)
 
 ## [0.5.0] - 2023-11-01